requirements.txt
colorama~=0.4.6
python-telegram-bot~=21.7
python-dotenv~=1.0.1
pytube~=15.0.0
yt-dlp~=2024.11.18
deep-translator~=1.11.4
google~=3.0.0
protobuf~=5.29.1